v0.0.0.1 - Creating basic structures
This time we started to create the basic files of our system, first we need to separate and component the engine into two modules, the game module and the editing module. In the game module we are going to define the configurations of the code and command interpreter of the game execution, in the edition module it is the engine functionality module in edition mode where we configure the game.