v0.0.0.7 -  Configuration files
This time we created some more files referring to the translation of the system for three languages, Portuguese, English and Spanish, these will be the three languages ​​we will use to translate the system at first, in the future as the project develops we will add other languages for translation. We also created some other classes to handle the translation.